blockchain_ingestion/requirements.txt
web3==6.11.0
pandas==2.0.3
pyyaml==6.0.1
asyncio
dataclasses
logging
datetime
typing